refactor: move selection sync logic into a hook
