+++ packages/elements-react/lib/components/Lens/hooks/index.ts
@@ -6,6 +6,7 @@ export { useFetchMoreOnScroll } from "./src/useFetchMoreOnScroll";
 export { useInfiniteFetch } from "./src/useInfiniteFetch";
 export { useColumnResize } from "./src/useColumnResize";
 export { useRowVirtualizer } from "./src/useRowVirtualizer";
+export { useSyncSelection } from "./src/useSyncSelection";
 
 export type { ResizeObserverCallback } from "./src/useResizeObserver";
 export type { UseInfiniteFetchParams } from "./src/useInfiniteFetch";
