+++ packages/elements-react/lib/components/Lens/hooks/src/useSyncSelection.ts
@@ -1,0 +1,34 @@
+import { useEffect } from "react";
+
+const useSyncSelection = (
+  selectedItems: any,
+  rowSelection: any,
+  setRowSelection: React.Dispatch<React.SetStateAction<any>>,
+  onSelect: (selectedIds: string[]) => void
+) => {
+  useEffect(() => {
+    const getOrderedKeys = (obj: any) => Object.keys(obj).sort().join(",");
+
+    const selectedItemsAreDifferentFromSelectedRows =
+      getOrderedKeys(selectedItems) !== getOrderedKeys(rowSelection);
+
+    if (selectedItemsAreDifferentFromSelectedRows) {
+      setRowSelection(selectedItems);
+    }
+  }, [selectedItems, rowSelection, setRowSelection]);
+
+  useEffect(() => {
+    const getOrderedKeys = (obj: any) => Object.keys(obj).sort().join(",");
+
+    const selectedItemsAreDifferentFromSelectedRows =
+      getOrderedKeys(selectedItems) !== getOrderedKeys(rowSelection);
+
+    if (selectedItemsAreDifferentFromSelectedRows) {
+      onSelect(Object.keys(rowSelection));
+    }
+  }, [rowSelection, selectedItems, onSelect]);
+};
+
+useSyncSelection.displayName = "useSyncSelection";
+
+export { useSyncSelection };
